Create initial her-gym communication layer
This commit creates server.py which can process all the commands required
for demo.py. It also creates client.py which manages the requests for
all the commands. I've also updated demo.py to use the connection
through client instead of gym directly.

The reason to do all this is to prevent the expensive gym
rendering/processing from operating on the jetson, removing processing
power from what we are actually trying to benchmark. Currently only
demo.py works with the server version of gym/mujoco, but we will have to
update train.py to use this connection too.