Create initial dockerfile
This dockerfile installs the basics in order to get the repo to work.
It doesn't install mujoco since that will be run on the server. The
docker container has been tested (and works) with demo.py on the
server-communication branch.

Instructions on how to create and run the docker container are at the
top of the dockerfile.