Add support for longer messages from client.py
Previously the server would just receive a set number of bytes and use
that data. However this has issues when longer sets of data are sent,
and if the data isn't received all at once. This was especially an issue
for COMPUTE_REWARD_CMD since it sends lots of data.